server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.21.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.8
cobble==0.1.4
coloredlogs==15.0.1
composio==0.1.1
composio_core==0.7.15
composio_langchain==0.7.15
cryptography==44.0.2
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
docker==7.1.0
docx2txt==0.8
durationpy==0.9
dynaconf==3.2.10
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperdock-container==0.5.7
hyperdock-langchain==0.5.7
hyperpocket==0.5.7
hyperpocket-langchain==0.5.7
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==32.0.1
langchain==0.3.24
langchain-chroma==0.2.3
langchain-community==0.3.23
langchain-core==0.3.56
langchain-openai==0.3.8
langchain-text-splitters==0.3.8
langchainhub==0.1.21
langgraph==0.4.1
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langgraph-supervisor==0.0.20
langsmith==0.3.39
lxml==5.4.0
magika==0.6.2
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
multiprocess==0.70.18
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.21.1
openai==1.76.2
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
ormsgpack==1.9.1
overrides==7.7.0
packaging==23.2
pandas==2.2.3
paramiko==3.5.1
pdfminer.six==20250416
pica-langchain==1.2.6
pillow==11.2.1
posthog==3.25.0
propcache==0.3.1
protobuf==5.29.4
pulsar-client==3.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydub==0.25.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pypdf==3.17.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
Pysher==1.0.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
semver==3.0.4
sentry-sdk==2.27.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tqdm==4.67.1
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
youtube-transcript-api==1.0.3
zipp==3.21.0
zstandard==0.23.0
